Move to correct file and fix ordering
